# src/NSwag.Sample.NETCore20/Output/swagger_new_v2.json
{
  "x-generator": "NSwag v11.18.3.0 (NJsonSchema v9.10.67.0 (Newtonsoft.Json v10.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "host": "localhost:65384",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json",
    "text/plain",
    "text/json"
  ],
  "paths": {
    "/pet": {
      "post": {
        "tags": [
          "Pet"
        ],
        "operationId": "Pet_AddPet",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "pet",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Pet"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SerializableError"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pet"
        ],
        "operationId": "Pet_EditPet",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "pet",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Pet"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SerializableError"
            }
          }
        }
      }
    },
    "/pet/findByStatus": {
      "get": {
        "tags": [
          "Pet"
        ],
        "operationId": "Pet_FindByStatus",
        "parameters": [
          {
            "type": "array",
            "name": "status",
            "in": "query",
            "required": true,
            "collectionFormat": "multi",
            "x-nullable": true,
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pet"
              }
            }
          },
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SerializableError"
            }
          }
        }
      }
    },
    "/pet/findByCategory": {
      "get": {
        "tags": [
          "Pet"
        ],
        "operationId": "Pet_FindByCategory",
        "parameters": [
          {
            "type": "string",
            "name": "category",
            "in": "query",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pet"
              }
            }
          },
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SerializableError"
            }
          }
        }
      }
    },
    "/pet/{petId}": {
      "get": {
        "tags": [
          "Pet"
        ],
        "operationId": "Pet_FindById",
        "parameters": [
          {
            "type": "integer",
            "name": "petId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/Pet"
            }
          },
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SerializableError"
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Pet"
        ],
        "operationId": "Pet_EditPetWithFormData",
        "parameters": [
          {
            "type": "integer",
            "name": "petId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "Id",
            "in": "formData",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "Age",
            "in": "formData",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "Category.Id",
            "in": "formData",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "Category.Name",
            "in": "formData",
            "required": true,
            "x-nullable": true
          },
          {
            "type": "boolean",
            "name": "HasVaccinations",
            "in": "formData",
            "required": true,
            "x-nullable": false
          },
          {
            "type": "string",
            "name": "Name",
            "in": "formData",
            "required": true,
            "x-nullable": true
          },
          {
            "type": "array",
            "name": "Images",
            "in": "formData",
            "required": true,
            "collectionFormat": "multi",
            "x-nullable": true,
            "items": {
              "$ref": "#/definitions/Image"
            }
          },
          {
            "type": "array",
            "name": "Tags",
            "in": "formData",
            "required": true,
            "collectionFormat": "multi",
            "x-nullable": true,
            "items": {
              "$ref": "#/definitions/Tag"
            }
          },
          {
            "type": "string",
            "name": "Status",
            "in": "formData",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SerializableError"
            }
          },
          "404": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Pet"
        ],
        "operationId": "Pet_DeletePet",
        "parameters": [
          {
            "type": "integer",
            "name": "petId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SerializableError"
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/pet/{petId}/uploadImage": {
      "post": {
        "tags": [
          "Pet"
        ],
        "operationId": "Pet_UploadImage",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "type": "integer",
            "name": "petId",
            "in": "path",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "file",
            "name": "file",
            "in": "formData",
            "required": true,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "400": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/SerializableError"
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/pet/RequiredAndOptional": {
      "post": {
        "tags": [
          "Pet"
        ],
        "operationId": "Pet_RequiredAndOptional",
        "parameters": [
          {
            "type": "integer",
            "name": "int_RequiredAndNotNullable",
            "in": "query",
            "required": true,
            "format": "int32",
            "x-nullable": false
          },
          {
            "type": "integer",
            "name": "int_RequiredAndNullable",
            "in": "query",
            "required": true,
            "format": "int32",
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "string_RequiredAndNullable",
            "in": "query",
            "required": true,
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "string_RequiredAndNotNullable",
            "in": "query",
            "required": true,
            "x-nullable": false
          },
          {
            "type": "number",
            "name": "decimalWithDefault_NotRequiredAndNotNullable",
            "in": "query",
            "format": "decimal",
            "default": 1.0,
            "x-nullable": false
          },
          {
            "type": "number",
            "name": "decimalWithDefault_NotRequiredAndNullable",
            "in": "query",
            "format": "decimal",
            "default": 1.0,
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "stringWithDefault_NotRequiredAndNullable",
            "in": "query",
            "default": "foo",
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "stringWithDefault_NotRequiredAndNotNullable",
            "in": "query",
            "default": "foo",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/sample": {
      "post": {
        "tags": [
          "Sample"
        ],
        "operationId": "Sample_GetSample",
        "responses": {
          "200": {
            "x-nullable": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Pet": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id",
        "age",
        "hasVaccinations",
        "name",
        "status"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "age": {
          "type": "integer",
          "format": "int32",
          "maximum": 150.0,
          "minimum": 0.0
        },
        "category": {
          "$ref": "#/definitions/Category"
        },
        "hasVaccinations": {
          "type": "boolean"
        },
        "name": {
          "type": "string",
          "maxLength": 50,
          "minLength": 2
        },
        "images": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Image"
          }
        },
        "tags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Tag"
          }
        },
        "status": {
          "type": "string",
          "minLength": 1
        }
      }
    },
    "Category": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Image": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "url": {
          "type": "string"
        }
      }
    },
    "Tag": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "SerializableError": {
      "type": "object",
      "additionalProperties": false,
      "allOf": [
        {
          "type": "object",
          "additionalProperties": {}
        }
      ]
    },
    "ApiResponse": {
      "type": "object",
      "additionalProperties": false,
      "required": [
        "code"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "TEST_HEADER": {
      "type": "apiKey",
      "description": "TEST_HEADER",
      "name": "TEST_HEADER",
      "in": "header"
    }
  }
}